@@ -1,28 +1,34 @@
 <script>
   import { logout } from "./auth.mjs";
-  import { userStore, route } from "./stores.mjs";
+  import { userStore, route } from "./stores.svelte.js";
 
-  console.log($userStore);
+  $inspect(userStore);
 </script>
 
 <nav>
   <ul>
-    <li><a href="#home" class={$route == "#home" ? "active" : ""}>Home</a></li>
     <li>
-      <a href="#posts" class={$route == "#posts" ? "active" : ""}>Posts</a>
+      <a href="#home" class={route.pathname == "#home" ? "active" : ""}>Home</a>
     </li>
-    {#if $userStore.isLoggedIn === false}
+    <li>
+      <a href="#posts" class={route.pathname == "#posts" ? "active" : ""}
+        >Posts</a
+      >
+    </li>
+    {#if userStore.isLoggedIn === false}
       <li>
-        <a href="#login" class={$route == "#login" ? "active" : ""}>Login</a>
+        <a href="#login" class={route.pathname == "#login" ? "active" : ""}
+          >Login</a
+        >
       </li>
-    {:else if $userStore.isLoggedIn === true}
+    {:else if userStore.isLoggedIn === true}
       <li>
-        <a href="#profile" class={$route == "#profile" ? "active" : ""}
+        <a href="#profile" class={route.pathname == "#profile" ? "active" : ""}
           >User Profile</a
         >
       </li>
 
-      <li><a href="#home" on:click={logout}>Logout</a></li>
+      <li><a href="#home" onclick={logout}>Logout</a></li>
     {/if}
   </ul>
 </nav>
