@@ -1,8 +1,9 @@
-import './app.css'
-import App from './App.svelte'
+import { mount } from "svelte";
+import "./app.css";
+import App from "./App.svelte";
 
-const app = new App({
-  target: document.getElementById('app'),
-})
+const app = mount(App, {
+  target: document.getElementById("app")
+});
 
-export default app
+export default app;
